Refactor various input hook managers to ensure naming consistency
